Working on Filters Screen
Duplicated browse screen components to make filters screen. Buttons are selectable instead of navigable.
